requirements.txt
requests==2.32.3
pydantic==2.8.2
PyYAML==6.0.1
python-dotenv==1.0.1
CurrencyConverter==0.17.28